requirements.txt
beautifulsoup4==4.5.1
bs4==0.0.1
certifi==2016.9.26
DateTime==4.1.1
defusedxml==0.4.1
Django==1.10.4
django-addthis==3.0.0
django-admin-honeypot==1.0.0
django-allauth==0.28.0
django-bootstrap3==7.1.0
django-ckeditor==5.1.1
django-contrib-comments==1.7.3
django-crispy-forms==1.6.1
django-djconfig==0.5.0
django-el-pagination==3.0.1
django-filter==0.15.3
django-fluent-comments==1.2.2
django-grappelli==2.9.1
django-haystack==2.5.1
django-infinite-scroll-pagination==0.1.3
django-mathfilters==0.4.0
django-multiupload==0.5.2
django-postman==3.4.0
django-relationships==0.3.3
django-smart-selects==1.2.6
django-tag-parser==3.0
django-threadedcomments==1.0.1
django-tinymce==2.4.0
elasticsearch==5.0.1
mistune==0.7.3
oauthlib==2.0.0
pbr==1.10.0
Pillow==3.4.2
psycopg2==2.6.1
pyelasticsearch==1.4
pysolr==3.6.0
python3-openid==3.0.10
pytils==0.3
pytz==2016.7
requests==2.11.1
requests-oauthlib==0.7.0
simplejson==3.10.0
six==1.10.0
sorl-thumbnail==12.3
stevedore==1.18.0
unicode-slugify==0.1.3
Unidecode==0.4.19
urllib3==1.19
URLObject==2.4.2
virtualenv==15.0.3
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.2
Whoosh==2.7.4
xlwt==1.1.2
zope.interface==4.3.2
